refactor(rust): extract sessions.rs with session discovery commands
